requirement.txt
streamlit==1.35.0
python-dotenv==1.0.1
google-generativeai==0.5.4
chromadb==0.5.0
langchain-chroma==0.1.1